Tegu Housekeeping - I fix my own tcomms mistake that I gave jegorg (#4122)
## About The Pull Request

Fixes some issues on the Tegu and adds a few things that were brought up
by people that played on it.

**Fixes the Tegu's tcomms to not link with the outpost and autolink with
other NT ships.**

Adds a single pair of sunglasses to the security officer's closet.
Adds a cigarette vendor in the hallway.
Adds two noticeboards to the vessel, for use in organization and
notice-boarding. This also gives vessel captains the ability to prevent
violations of OSHA code [1910 subpart
E](https://www.osha.gov/laws-regs/regulations/standardnumber/1910/1910SubpartE).
Adds a PDA painter to customize PDAs the vessel is given.
Unifies the Paperwork Storage area with the Supply Office area to reduce
power usage on the ship.
Gives the engineer an angle grinder to assist in salvaging and mining
operations.
Adds two fire cabinets to bring the vessel in compliance with OSHA code
[1915.505](https://www.osha.gov/laws-regs/interlinking/standards/1915.505).
Adds PDA cartridges to cabinets for use with the jobs' respective PDAs
(if they are chosen to be used).
Stocks the oxygen cabinets with two small emergency oxygen tanks as they
did not spawn with oxygen bottles to begin with.
Adds NT and N+S stamps where they belong.

## Why It's Good For The Game

This ship had a number of issues and was missing a few things, this PR
seeks to alleviate those issues in a non-intrusive manner.
The issue of the Tegu linking with every single relay and ship has been
resolved as well.

A few gear additions help make the Tegu lean more towards it's
salvage/mining capacity and give the Engineer the option of fieldwork
assistance.

## Changelog

:cl:
balance: A number of changes have been made to the Tegu to modernize it,
including the addition of an angle grinder and a repaired
telecommunications array.
/:cl: