Minor Sunskipper upkeep. Also gives it a rifle. (#4147)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
- Adds a Model-12 Sporter because I think the ship can use a hunting
rifle (even if it shoots .22)
- The safe has a written-down code now, instead of you having to use the
stethoscope
- Fixes colors of railings and stuff
- Other minor tweaks and fixes

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Maintaining my ship I oh so hate
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Changelog

:cl:
add: Added a Model 12 Sporter to the Sunskipper.
tweak: Sunskipper's safe now has the combination written down on a piece
of paper - no longer do you need the stethoscope.
fix: Fixed colors of Sunskipper's objects.
tweak: A bunch of really minor tweaks and fixes to the Sunskipper.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->