Plexagon ID management can assign job's from the ship's job list. (#4128)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

- If a console get started on a ship it will now be able to assign an ID
to any job that can be joined on the ship.
- This behavior can be extended to any computer if you copy Plexagon
acces management to another computer or yoink the hard drive, but you'll
still need a card with proper access (including ship access) to do so.

The old behavior is still present if Plexagon is not connected to any
ship.

## Why It's Good For The Game

Id management felt off since we were only able to assign crew to station
job. With this pr we will now be able to actualy demote/promote crew.

## Changelog

:cl:
add: Assigning a crew to a job that comes on the ship is now possible
with command consoles.
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: ritorizo <ritorizo@localhost>