Scrapping Charges in Cargo (#4146)
## About The Pull Request

Adds a C-4 duffel to cargo (10 charges at 100 a charge), at the cost of
1000 credits.
C-4 itself is surprisingly not as destructive as one might think, and in
the things one performs in shiptest it could be particularly useful.
Breaching a wall in a ruin is a big one, or blowing open safes (did you
know you could do that with C4?)

C-4 is already available in a few places and even on certain military
ships roundstart, so it wouldn't be too out of place.

This PR also makes it so you only need one C-4 charge to blow the lock
on a safe, to make them another method to bypass the lock. Three charges
on what was effectively rare to find explosives just to pop open a safe
was a bit much. (Refactoring was done by @FalloutFalcon, thanks
Fallcon!)

## Why It's Good For The Game

This adds additional gameplay options for clearing ruins, excavating,
breaking open safes, and a number of other things. With a do_after to
place them, a 10 second timer and a very visible overlay on items, it's
hard to use them stealthily either.

## Changelog

:cl: Cloudbreak, FalloutFalcon
add: Demolitions charges to Cargo
balance: Safes only take one C-4 charge to explode
/:cl: